ssl: splitting socket lib into client and server libraries
