+++ mobile/test/performance/files_em_does_not_use
@@ -11,3 +11,4 @@ source/extensions/access_loggers/common/file_access_log_impl.h
 source/common/router/scoped_rds.h
 source/extensions/load_balancing_policies/subset/subset_lb.h
 source/extensions/transport_sockets/tls/downstream_config.h
+source/common/tls/server_ssl_socket.h
