+++ source/common/quic/BUILD
@@ -181,7 +181,7 @@ envoy_cc_library(
         "//envoy/http:persistent_quic_info_interface",
         "//envoy/registry",
         "//source/common/runtime:runtime_lib",
-        "//source/common/tls:ssl_socket_lib",
+        "//source/common/tls:client_ssl_socket_lib",
         "//source/extensions/quic/crypto_stream:envoy_quic_crypto_client_stream_lib",
         "@com_github_google_quiche//:quic_core_http_spdy_session_lib",
     ],
@@ -462,8 +462,8 @@ envoy_cc_library(
         "//envoy/ssl:context_config_interface",
         "//source/common/common:assert_lib",
         "//source/common/network:transport_socket_options_lib",
+        "//source/common/tls:client_ssl_socket_lib",
         "//source/common/tls:context_config_lib",
-        "//source/common/tls:ssl_socket_lib",
         "@com_github_google_quiche//:quic_core_crypto_crypto_handshake_lib",
         "@envoy_api//envoy/extensions/transport_sockets/quic/v3:pkg_cc_proto",
     ],
@@ -488,7 +488,7 @@ envoy_cc_library(
         "//source/common/common:assert_lib",
         "//source/common/network:transport_socket_options_lib",
         "//source/common/tls:context_config_lib",
-        "//source/common/tls:ssl_socket_lib",
+        "//source/common/tls:server_ssl_socket_lib",
         "@com_github_google_quiche//:quic_core_crypto_crypto_handshake_lib",
         "@envoy_api//envoy/extensions/transport_sockets/quic/v3:pkg_cc_proto",
     ],
