+++ test/common/grpc/grpc_client_integration_test_harness.h
@@ -27,7 +27,8 @@
 #include "source/common/stats/symbol_table.h"
 
 #include "source/common/tls/context_config_impl.h"
-#include "source/common/tls/ssl_socket.h"
+#include "source/common/tls/client_ssl_socket.h"
+#include "source/common/tls/server_ssl_socket.h"
 
 #include "test/common/grpc/grpc_client_integration.h"
 #include "test/common/grpc/utility.h"
