+++ test/extensions/transport_sockets/starttls/upstream_starttls_integration_test.cc
@@ -3,7 +3,7 @@
 
 #include "source/common/network/connection_impl.h"
 #include "source/common/tls/context_config_impl.h"
-#include "source/common/tls/ssl_socket.h"
+#include "source/common/tls/server_ssl_socket.h"
 #include "source/extensions/filters/network/common/factory_base.h"
 #include "source/extensions/transport_sockets/starttls/starttls_socket.h"
