+++ test/integration/alpn_selection_integration_test.cc
@@ -5,7 +5,7 @@
 #include "source/common/http/utility.h"
 #include "source/common/tls/context_config_impl.h"
 #include "source/common/tls/context_impl.h"
-#include "source/common/tls/ssl_socket.h"
+#include "source/common/tls/server_ssl_socket.h"
 
 #include "test/integration/http_integration.h"
