+++ test/integration/base_integration_test.cc
@@ -19,7 +19,7 @@
 #include "source/common/event/libevent.h"
 #include "source/common/network/utility.h"
 #include "source/common/tls/context_config_impl.h"
-#include "source/common/tls/ssl_socket.h"
+#include "source/common/tls/server_ssl_socket.h"
 #include "source/server/proto_descriptors.h"
 
 #include "test/integration/utility.h"
