+++ test/integration/sds_dynamic_integration_test.cc
@@ -21,7 +21,7 @@
 
 #include "source/common/tls/context_config_impl.h"
 #include "source/common/tls/context_manager_impl.h"
-#include "source/common/tls/ssl_socket.h"
+#include "source/common/tls/server_ssl_socket.h"
 
 #include "test/common/grpc/grpc_client_integration.h"
 #include "test/config/integration/certs/clientcert_hash.h"
